Multi-Agent-Pipeline/requirements.txt
# Multi-Agent Deploy Pipeline: Terraform → Build → Deploy → Verify
# Pin these to avoid slow "pip is looking at multiple versions" resolution on install.
instructor==1.12.0
opentelemetry-exporter-otlp-proto-grpc==1.34.1
crewai>=0.80.0
crewai-tools>=0.14.0
requests>=2.28.0
boto3>=1.26.0
python-dotenv>=1.0.0
